requirements.txt
anyio==4.3.0
blinker==1.7.0
bottle==0.12.25
cffi==1.16.0
click==8.1.7
clr-loader==0.2.6
colorama==0.4.6
Faker==24.4.0
Flask==3.0.2
flask-peewee==3.0.4
idna==3.6
itsdangerous==2.1.2
Jinja2==3.1.3
MarkupSafe==2.1.5
peewee==3.17.1
proxy_tools==0.1.0
pycparser==2.21
python-dateutil==2.9.0.post0
pythonnet==3.0.3
pywebview==5.0.5
six==1.16.0
sniffio==1.3.1
typing_extensions==4.10.0
watchfiles==0.21.0
Werkzeug==3.0.1
wtf-peewee==3.0.5
WTForms==2.3.0
